# api/docs/swagger.yaml
definitions:
  models.BusCreate:
    properties:
      destination_id:
        type: integer
      end_time:
        type: string
      name:
        type: string
      seat_count:
        type: integer
      start_time:
        type: string
    type: object
  models.BusStopCreate:
    properties:
      destination_id:
        type: integer
      distance:
        type: integer
      name:
        type: string
    type: object
  models.DestinationCreate:
    properties:
      distance:
        type: integer
      from:
        type: string
      price:
        type: integer
      to:
        type: string
    type: object
  models.DestinationUpdate:
    properties:
      distance:
        type: integer
      from:
        type: string
      price:
        type: integer
      to:
        type: string
    type: object
  models.ResponseError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.SuccessModel:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/bus:
    post:
      consumes:
      - application/json
      description: API for create bus
      parameters:
      - description: bus
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.BusCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create bus
      tags:
      - bus
  /v1/bus-stop:
    post:
      consumes:
      - application/json
      description: API for create bus-stop
      parameters:
      - description: bus-stop
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.BusStopCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create bus-stop
      tags:
      - bus-stop
  /v1/bus-stop/{destination_id}:
    get:
      consumes:
      - application/json
      description: API for getting all bus-stopes
      parameters:
      - description: destination_id
        in: path
        name: destination_id
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: GetAll bus-stop
      tags:
      - bus-stop
  /v1/bus-stop/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting bus-stop
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete bus-stop
      tags:
      - bus-stop
  /v1/destination:
    get:
      consumes:
      - application/json
      description: API for getting all destinations
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: GetAll destinations
      tags:
      - destination
    post:
      consumes:
      - application/json
      description: API for create destination
      parameters:
      - description: destination
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.DestinationCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create destination
      tags:
      - destination
  /v1/destination/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting destination
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete destination
      tags:
      - destination
    get:
      consumes:
      - application/json
      description: API for getting destination
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get destination
      tags:
      - destination
    put:
      consumes:
      - application/json
      description: API for updating destination
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: destination
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.DestinationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update destination
      tags:
      - destination
swagger: "2.0"