# api/models/bus.go
package models

type BusCreate struct {
	Name          string `json:"name" db:"name"`
	DestinationID int64  `json:"destination_id" db:"destination_id"`
	StartTime     string `json:"start_time" db:"start_time"`
	EndTime       string `json:"end_time" db:"end_time"`
	SeatCount     int64  `json:"seat_count" db:"seat_count"`
}

type Bus struct {
	ID     int64 `json:"id" db:"id"`
	IsFull bool  `json:"is_full" db:"is_full"`
	BusCreate
}